Refactor schma inference to decouple from scraper
